Refactor more connection logic to base session
