+++ src/IFY.Phorm.SqlClient/SqlPhormSession.cs
@@ -2,30 +2,12 @@
 using IFY.Phorm.Execution;
 using Microsoft.Data.SqlClient;
 using System;
-using System.Collections.Generic;
 using System.Data;
 
 namespace IFY.Phorm.SqlClient
 {
     public class SqlPhormSession : AbstractPhormSession
     {
-        public event EventHandler<IPhormDbConnection>? Connected;
-
-        private static readonly Dictionary<string, IPhormDbConnection> _connectionPool = new Dictionary<string, IPhormDbConnection>();
-#if DEBUG
-        internal static void ResetConnectionPool()
-        {
-            lock (_connectionPool)
-            {
-                foreach (var conn in _connectionPool.Values)
-                {
-                    conn.Dispose();
-                }
-                _connectionPool.Clear();
-            }
-        }
-#endif
-
         internal Func<string, string?, IPhormDbConnection> _connectionBuilder = (sqlConnStr, connectionName) => new PhormDbConnection(connectionName, new SqlConnection(sqlConnStr));
 
         public SqlPhormSession(string databaseConnectionString, string? connectionName = null)
@@ -41,59 +23,29 @@ namespace IFY.Phorm.SqlClient
             };
         }
 
-        // TODO: base?
-        protected override IPhormDbConnection GetConnection()
+        protected override IPhormDbConnection CreateConnection()
         {
-            // Reuse existing connections, where possible
-            if (!_connectionPool.TryGetValue(ConnectionName ?? string.Empty, out var phormConn)
-                || phormConn.State != ConnectionState.Open)
+            // Ensure application name is known user
+            var connectionString = new SqlConnectionStringBuilder(_databaseConnectionString);
+            connectionString.ApplicationName = ConnectionName ?? connectionString.ApplicationName;
+            var sqlConnStr = connectionString.ToString();
+
+            // Create connection
+            var conn = _connectionBuilder(sqlConnStr, ConnectionName);
+            if (conn.DefaultSchema.Length == 0)
             {
-                lock (_connectionPool)
-                {
-                    if (!_connectionPool.TryGetValue(ConnectionName ?? string.Empty, out phormConn)
-                        || phormConn.State != ConnectionState.Open)
-                    {
-                        // Create new connection
-                        phormConn?.Dispose();
-
-                        // Ensure application name is known user
-                        var connectionString = new SqlConnectionStringBuilder(_databaseConnectionString);
-                        connectionString.ApplicationName = ConnectionName ?? connectionString.ApplicationName;
-                        var sqlConnStr = connectionString.ToString();
-
-                        // Create connection
-                        phormConn = _connectionBuilder(sqlConnStr, ConnectionName);
-
-                        // Resolve default schema
-                        if (phormConn.DefaultSchema.Length == 0)
-                        {
-                            using var cmd = ((IDbConnection)phormConn).CreateCommand();
-                            cmd.CommandText = "SELECT schema_name()";
-                            phormConn.DefaultSchema = cmd.ExecuteScalar()?.ToString() ?? connectionString.UserID;
-                        }
-                        _connectionPool[ConnectionName ?? string.Empty] = phormConn;
-
-                        try
-                        {
-                            Connected?.Invoke(this, phormConn);
-                        }
-                        catch { }
-                    }
-                }
+                conn.DefaultSchema = connectionString.UserID;
             }
-            return phormConn;
+            return conn;
         }
 
         #region Console capture
 
         protected override AbstractConsoleMessageCapture StartConsoleCapture(Guid commandGuid, IAsyncDbCommand cmd)
         {
-            if (cmd.Connection is SqlConnection sql)
-            {
-                return new SqlConsoleMessageCapture(this, commandGuid, sql);
-            }
-
-            return NullConsoleMessageCapture.Instance;
+            return cmd.Connection is SqlConnection sql
+                ? new SqlConsoleMessageCapture(this, commandGuid, sql)
+                : (AbstractConsoleMessageCapture)NullConsoleMessageCapture.Instance;
         }
 
         #endregion Console capture
