+++ src/IFY.Phorm.Tests/Execution/AbstractPhormSessionTests.cs
@@ -25,6 +25,12 @@ namespace IFY.Phorm.Tests
         {
         }
 
+        [TestInitialize]
+        public void Init()
+        {
+            AbstractPhormSession.ResetConnectionPool();
+        }
+
         [TestMethod]
         public void Defaults_prefixes_to_GlobalSettings()
         {
